% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_albers.R
\name{albers_complement}
\alias{albers_complement}
\title{Return a complementary family for diverging palettes}
\usage{
albers_complement(
  family = c("red", "lapis", "ochre", "teal", "green", "violet")
)
}
\arguments{
\item{family}{One of "red","lapis","ochre","teal","green","violet"}
}
\description{
Pairs warm/cool and related families to produce balanced diverging
combinations that align with the Homage system.
}
\keyword{internal}
