% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_img.R
\name{scale_color_albers_img}
\alias{scale_color_albers_img}
\alias{scale_fill_albers_img}
\title{Image-derived sequential color scale}
\usage{
scale_color_albers_img(family = "red", discrete = TRUE, ...)

scale_fill_albers_img(family = "red", discrete = TRUE, ...)
}
\arguments{
\item{family}{One of "red","lapis","ochre","teal","green"}

\item{discrete}{Whether to use discrete palette; if FALSE uses a gradient.}

\item{...}{Passed to underlying ggplot2 scale.}
}
\value{
A \code{ggplot2} scale object.
}
\description{
Image-derived sequential color scale
}
