% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_albers.R
\name{scale_linetype_albers}
\alias{scale_linetype_albers}
\title{Discrete linetype scale to pair with Albers colors}
\usage{
scale_linetype_albers(...)
}
\arguments{
\item{...}{Passed to \code{ggplot2::scale_linetype_manual()}.}
}
\description{
Provides a sensible set of linetypes for multi-series line charts.
}
