% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_albers.R
\name{gt_albers}
\alias{gt_albers}
\title{Quiet, legible gt style with subtle stripe from A300}
\usage{
gt_albers(x, family = "red")
}
\arguments{
\item{x}{A \code{gt} table}

\item{family}{Palette family for subtle accents}
}
\description{
Quiet, legible gt style with subtle stripe from A300
}
