% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_albers.R
\name{use_albers_vignettes}
\alias{use_albers_vignettes}
\title{Configure current package to use albersdown}
\usage{
use_albers_vignettes()
}
\description{
\itemize{
\item Copies vignette CSS if missing
\item Suggests helper packages
\item Adds pkgdown template stanza to \verb{_pkgdown.yml}
}
}
