% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_albers.R
\name{use_albers_vignettes}
\alias{use_albers_vignettes}
\title{Configure current package to use albersdown (back-compat wrapper)}
\usage{
use_albers_vignettes()
}
\description{
This wrapper preserves the old name and forwards to \code{use_albersdown()}.
}
